% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopulationWaves.R
\name{getPopulationWaves}
\alias{getPopulationWaves}
\title{getPopulationWaves: Visualize celltypes along pseudotemporal coordinates}
\usage{
getPopulationWaves(
  sce,
  pseudotime_sce,
  celltype,
  plot_as_function_values,
  population_smooth_factor = 0.5,
  matlab_version,
  cyt3_script = "~/Documents/MATLAB",
  output_dir = "~/Downloads"
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing selected cells from the isolated trajectory with scaled pseudotemporal coordinates.}

\item{pseudotime_sce}{A column of numeric pseudotemporal coordinate values in SingleCellExperiment object.}

\item{celltype}{A column of character values representing celltypes within the trajectory.}

\item{plot_as_function_values}{A data frame containing: pseudotemporal locations, features, and smoothed expression values.}

\item{population_smooth_factor}{Smooth factor to visualize population peaks along the trajectory. Smaller values gives sharper populations peaks and narrower transitions. Larger values gives wider populations peaks and smoother transitions. Default to 0.5.}

\item{matlab_version}{A character string of active Matlab version (e.g. "R2026a").}

\item{cyt3_script}{Path to Matlab folder containing cyt3 scripts. Default to "~/Documents/MATLAB".}

\item{output_dir}{Path to store output files from running plot_as_function. Default to "~/Downloads".}
}
\value{
Attaches columns containing celltype coordinates to plot_as_function data frame.
}
\description{
This function isolates each defined celltype along pseudotemporal coordinates to visualize their location
and gene expression(s) along a trajectory. Celltype coordinates are applied gaussian smoothing and individually rescaled
to follow expression values along runPlotFun()'s pseudotemporal locations.
Must first run runPlotFun() to run this function.
}
\author{
Luis D. Alvarez
}
