% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlotFun.R
\name{runPlotFun}
\alias{runPlotFun}
\title{runPlotFun: Run plot_as_function.m algorithm.}
\usage{
runPlotFun(
  sce,
  pseudotime,
  selected_features,
  num_locs = 100L,
  smooth_factor = 0.5,
  assay = "logcounts",
  matlab_version,
  cyt3_script = "~/Documents/MATLAB",
  output_dir = "~/Downloads"
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing selected cells from the isolated trajectory.}

\item{pseudotime}{A numeric vector of pseudotime values, corresponds to cells' spatial position along the pseudotime.}

\item{selected_features}{A character vector of features to apply the Gaussian kernel (per gene) following the expression in pseudotime.}

\item{num_locs}{Number of pseudotime locations to represent gene expression values across pseudotime range. Every location calculates a Gaussian-weighted average of nearby cells' normalized expression. Defaults to 100L.}

\item{smooth_factor}{Width of the Gaussian neighborhood used to estimate expression at each pseudotime location. Smaller values gives sharper peaks and narrower transitions. Larger values gives wider peaks and smoother transitions. Defaults to 0.5.}

\item{assay}{Assay name containing normalized data in SingleCellExperiment object. Defaults to "logcounts".}

\item{matlab_version}{A character string of active Matlab version (e.g. "R2026a").}

\item{cyt3_script}{Path to Matlab folder containing Cyt3 scripts. Defaults to "~/Documents/MATLAB".}

\item{output_dir}{Path to store output files from running plot_as_function. Defaults to "~/Downloads".}
}
\value{
Input CSV file containing the pseudotime and normalized gene expression values that was used as input for plot_as.function.m in Matlab. Output CSV file containing pseudotime locations and smoothed gene expression values. Log file of plot_as_function.m output in Matlab.
}
\description{
User interface to run plot_as_function.m from cyt (Pe'er Lab) in Matlab. Function applies
a Gaussian kernel (per gene) on normalized values along a defined pseudotime trajectory.
Function modifies plot_as_function.m to handle deprecated functions and errors
to run on current Matlab versions.
}
\author{
Luis D. Alvarez
}
