% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlotFun.R
\name{runPlotFun}
\alias{runPlotFun}
\title{runPlotFun wrapper: Run plot_as_function algorithm to apply gaussian smoothing on log-normalized single cell data}
\usage{
runPlotFun(
  sce,
  pseudotime,
  selected_features,
  num_locs = 100L,
  smooth_factor = 0.5,
  assay = "logcounts",
  matlab_version,
  cyt3_script = "~/Documents/MATLAB",
  output_dir = "~/Downloads"
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing selected cells from the isolated trajectory with scaled pseudotemporal coordinates.}

\item{pseudotime}{A column of numeric pseudotemporal coordinate values in SingleCellExperiment object.}

\item{selected_features}{A character vector of features to apply smoothing.}

\item{num_locs}{Number of locations to represent expression values across pseudotemporal coordinates. Every location calculates a Gaussian-weighted average of nearby cells' normalized expression. Default to 100L.}

\item{smooth_factor}{Width of the Gaussian neighborhood used to estimate expression at each location. Smaller values gives sharper peaks and narrower transitions. Larger values gives wider peaks and smoother transitions. Default to 0.5.}

\item{assay}{Assay name containing normalized data in SingleCellExperiment object. Default to "logcounts".}

\item{matlab_version}{A character string of active Matlab version (e.g. "R2026a").}

\item{cyt3_script}{Path to Matlab folder containing cyt3 scripts. Default to "~/Documents/MATLAB".}

\item{output_dir}{Path to store output files from running plot_as_function. Default to "~/Downloads".}
}
\value{
Input CSV file containing the pseudotime and normalized gene expression values that was used as input for plot_as.function.m in Matlab. Output CSV file containing pseudotime locations and smoothed gene expression values. Log file of plot_as_function.m console output in Matlab.
}
\description{
User interface to run plot_as_function.m from cyt (Pe'er Lab) in Matlab. Function applies
Gaussian smoothing (per gene) on normalized values along scaled pseudotemporal coordinates.
Function modifies plot_as_function.m to handle deprecated functions and errors
to run on current Matlab versions.
}
\author{
Luis D. Alvarez
}
